requirements.rag.txt
streamlit 
langchain 
langchain-community
ollama
openai 
faiss-cpu 
tiktoken 
chromadb 
pypdf
python-docx
pandas 
openpyxl


